Remove executable. Move Lib to Spec
