--- app/Main.hs
@@ -1,6 +1,0 @@
-module Main where
-
-import Lib
-
-main :: IO ()
-main = someFunc
