+++ dbrecord.cabal
@@ -16,7 +16,7 @@ cabal-version:       >=1.10
 library
   hs-source-dirs:      src
   ghc-options:         -Wall
-  exposed-modules:     Lib
+  exposed-modules:
                        Database.Migration
                        Database.Transaction
                        Database.Query
@@ -36,19 +36,12 @@ library
                      , attoparsec
   default-language:    Haskell2010
 
-executable dbrecord-exe
-  hs-source-dirs:      app
-  main-is:             Main.hs
-  ghc-options:         -threaded -rtsopts -with-rtsopts=-N -Wall
-  build-depends:       base
-                     , dbrecord
-  default-language:    Haskell2010
-
 test-suite dbrecord-test
   type:                exitcode-stdio-1.0
   hs-source-dirs:      test
   main-is:             Spec.hs
   build-depends:       base
+                     , text
                      , dbrecord
   ghc-options:         -threaded -rtsopts -with-rtsopts=-N -Wall
   default-language:    Haskell2010
